Move Conn interface to a separate conn.go
